wip: return proper message for invalid transactions

+++ src/models/event.py
@@ -3,6 +3,6 @@ from pydantic import BaseModel, Field
 
 class EventSchema(BaseModel):
     type: str = Field(...)
-    amount: int = Field(...)
+    amount: float = Field(...)
     destination: str = None
     origin: str = None
